Phase 1 (#358): rename connect/ → connector/ and lift capability-agnostic files to umbrella root
Mechanical refactor with no behavior change, setting up the layout for
the new importer/ subpackage that arrives in Phase 2.

The connect/ subpackage is renamed to connector/ for symmetry with
the forthcoming importer/. Five files that were already capability-
agnostic are lifted from connector/ up to the integrations/ umbrella
root so neither capability subpackage will depend on the other:

  - placement_request.py
  - integration_attribute_edit_context.py
  - entity_operations.py
  - integration_data.py
  - integration_gateway.py

The remaining connector/ contents (views, synchronizer, monitors,
sync_check, sync_result, view_mixins, external_view_data, plus the
Connect-mode controller/manage-view-pane stubs and user_data_detector)
stay capability-specific.

view_mixins.py stays in connector/ as a single file; a future
refactor could split the capability-agnostic accessor methods up to
umbrella, but the call-site benefit is small and bundling that with
this rename would obscure the move.

The defensive Disable-side data_source predicate originally listed
in Phase 1 is deferred to Phase 7. The naive EXTERNAL filter would
break HA/ZM/Frigate Disable (their entities are INTERNAL), so the
guard needs to be conditional on the integration declaring both
capabilities — designing that alongside Phase 7's symmetric block-
modal capability-conditional detection.
