+++ src/hi/integrations/connector/integration_manage_view_pane.py
