+++ src/hi/integrations/connector/sync_result.py
