+++ src/hi/integrations/tests/test_sync_check.py
@@ -15,13 +15,13 @@ from unittest.mock import Mock, patch
 from django.core.cache import cache
 from django.test import TestCase
 
-from hi.integrations.connect.sync_check import (
+from hi.integrations.connector.sync_check import (
     IntegrationSyncCheck,
     SyncCheckOutcome,
     SyncCheckResult,
     SyncDelta,
 )
-from hi.integrations.connect.monitors import IntegrationSyncCheckMonitor
+from hi.integrations.connector.monitors import IntegrationSyncCheckMonitor
 from hi.integrations.transient_models import IntegrationKey
 
 
@@ -658,7 +658,7 @@ class IntegrationSynchronizerPostSyncHookTests(TestCase):
         returns the given result. We test sync() (the public entry
         point) so the post-hook fires the same way it does in
         production."""
-        from hi.integrations.connect.integration_synchronizer import IntegrationSynchronizer
+        from hi.integrations.connector.integration_synchronizer import IntegrationSynchronizer
         from hi.integrations.transient_models import IntegrationMetaData
 
         class _TestSynchronizer(IntegrationSynchronizer):
@@ -678,7 +678,7 @@ class IntegrationSynchronizerPostSyncHookTests(TestCase):
         return _TestSynchronizer()
 
     def test_successful_sync_records_zero_delta_completion(self):
-        from hi.integrations.connect.sync_result import IntegrationSyncResult
+        from hi.integrations.connector.sync_result import IntegrationSyncResult
 
         # Pre-populate a stale "needs sync" state.
         IntegrationSyncCheck.set_state(
@@ -699,7 +699,7 @@ class IntegrationSynchronizerPostSyncHookTests(TestCase):
         self.assertFalse(loaded.needs_sync)
 
     def test_failed_sync_leaves_cache_alone(self):
-        from hi.integrations.connect.sync_result import IntegrationSyncResult
+        from hi.integrations.connector.sync_result import IntegrationSyncResult
 
         # Pre-populate a stale needs-sync state.
         stale = IntegrationSyncCheck.build_result(
