+++ src/hi/integrations/urls.py
@@ -4,7 +4,7 @@ from django.urls import re_path, include
 
 from hi.apps.common.module_utils import import_module_safe
 
-from .connect import views
+from .connector import views
 
 
 urlpatterns = [
