+++ src/hi/services/homebox/tests/test_hb_connector.py
@@ -13,7 +13,7 @@ from unittest.mock import Mock, patch
 from django.test import TestCase
 
 from hi.apps.entity.models import Entity
-from hi.integrations.connect.external_view_data import (
+from hi.integrations.connector.external_view_data import (
     MinimalViewData,
     NameValuePair,
     StructuredViewData,
