+++ src/hi/services/zoneminder/zm_sync.py
@@ -17,9 +17,9 @@ from hi.apps.entity.entity_placement import (
     EntityPlacementGroup,
 )
 
-from hi.integrations.connect.integration_synchronizer import IntegrationSynchronizer
-from hi.integrations.connect.sync_check import IntegrationSyncCheck, SyncDelta
-from hi.integrations.connect.sync_result import IntegrationSyncResult
+from hi.integrations.connector.integration_synchronizer import IntegrationSynchronizer
+from hi.integrations.connector.sync_check import IntegrationSyncCheck, SyncDelta
+from hi.integrations.connector.sync_result import IntegrationSyncResult
 from hi.integrations.transient_models import IntegrationKey
 
 from .zm_metadata import ZmMetaData
